% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.causalTree.R
\name{estimate.causalTree}
\alias{estimate.causalTree}
\title{Estimate the causal effects using honest tree model based on \code{causalTree} fit.}
\description{
Estimate the causal effects using honest tree model.
}
\usage{
estimate.causalTree(object, data, treatment, na.action = na.pass)
}
\arguments{
\item{object}{A tree-structured fit object, and here is specified as \code{causalTree} fit.}

\item{data}{New data frame in which to interact the variables named in the formula.}

\item{treatment}{The treatment status of new observations, 1 represents treated and 0 represents controlled.}
}

\details{The \code{estimate.causalTree} differs from \code{predict} in using the new data frame values to compute the causal effects. When the leaf contains only treated or control cases, the function will trace back to the leaf's parent node recursively until the parent can be used to compute causal effect.
Please check Susan and Guido's paper \emph{Machine Learning Methods for Estimating Heterogeneous Causal Effects} (2015) for detials.
}

\value{
a numeric vector indicating the estimated causal effects.
}

\references{
Susan A., Guido I. (2015)  \emph{Machine Learning Methods for Estimating Heterogeneous Causal Effects}.
}

\seealso{
  \code{\link{honest.causalTree}},
  \code{\link{causalTree}},
  \code{\link{predict.rpart}},
}

\examples{
fit <- causalTree(y~ x1 + x2 + x3 + x4, data = simulation.1, 
                  treatment = simulation.1$treatment, split.Rule = "CT", 
                  cv.option = "CT", split.Honest = T, 
                  cv.Honest = T, split.Bucket = F, xval = 5, 
                  cp = 0, minsize = 20, propensity = 0.5)
                  
opcp <- fit$cptable[,1][which.min(fit$cptable[,4])]

opfit <- prune(fit, opcp)


estimation <- estimate.causalTree(opfit, data = simulation.2, 
                                  treatment = simulation.2$treatment)
}