% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.distance.R
\name{get.tree.struct}
\alias{get.tree.struct}
\title{return the tree structure as grf}
\usage{
get.tree.struct(tree)
}
\arguments{
\item{tree}{the causal tree object we get from causalForest object}
}
\description{
return the tree structure as grf
}
