% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.distance.R
\name{get_node}
\alias{get_node}
\title{support function to get the node information}
\usage{
get_node(dep_ind, df)
}
\arguments{
\item{dep_ind}{the depth index of the node}

\item{df}{dataframe we get from first part of get.tree.struct function}
}
\description{
support function to get the node information
}
