% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.distance.R
\name{obtain_bs}
\alias{obtain_bs}
\title{Obtain binary strings.}
\usage{
obtain_bs(tree, data)
}
\arguments{
\item{tree}{casual tree object.}

\item{data}{the data used to train the casual forest.}
}
\description{
Obtain binary strings.
}
