% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.distance.R
\name{obtain_d2_distance}
\alias{obtain_d2_distance}
\title{Obtain d2 distance between two trees. (Deprecated)}
\usage{
obtain_d2_distance(tree1, tree2, data)
}
\arguments{
\item{tree1}{casual tree object 1.}

\item{tree2}{casual tree object 1.}

\item{data}{the data used to train the casual forest.}
}
\description{
Obtain d2 distance between two trees. (Deprecated)
}
