% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.distance.R
\name{obtain_hamming_distance}
\alias{obtain_hamming_distance}
\title{Obtain Hamming distance between two trees.}
\usage{
obtain_hamming_distance(binlist1, binlist2)
}
\arguments{
\item{binlist1}{binary strings corresponding to tree 1.}

\item{binlist2}{binary strings corresponding to tree 2.}
}
\description{
Obtain Hamming distance between two trees.
}
