% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.distance.R
\name{predict_tree}
\alias{predict_tree}
\title{Make prediction using single causal tree.}
\usage{
predict_tree(node, sample, tree)
}
\arguments{
\item{node}{node object in causal tree object.}

\item{sample}{single sample from the provided dataset (single patient).}

\item{tree}{Causal tree object.}
}
\description{
Make prediction using single causal tree.
}
