% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.causalTree.R
\name{estimate.causalTree}
\alias{estimate.causalTree}
\title{Estimate treatment effects within leaves of an \code{rpart} object, such as one generated by \code{causalTree}, using a new dataset.}
\description{
Estimate treatment effects within the leaves of a tree (a \code{rpart} object) using a new dataset.
}
\usage{
estimate.causalTree(object, data, treatment, na.action = na.pass)
}
\arguments{
\item{object}{A tree-structured fit \code{rpart} object, such as one generated as a \code{causalTree} fit.}

\item{data}{New data frame to be used for estimating effects within leaves.}

\item{treatment}{The treatment status of observations in the new dataframe, where 1 represents treated and 0 represents control.}
}

\details{When the leaf contains only treated or control cases, the function will trace back to the leaf's parent node recursively until the parent can be used to compute causal effect.
Please see Athey and Imbens \emph{Machine Learning Methods for Estimating Heterogeneous Causal Effects} (2015) for details.
}

\value{
An \code{rpart} object, where the leaf estimates are treatment effects within the leaves in the input dataset.
}

\references{
Susan A., Guido I. (2015)  \emph{Recursive Partitioning for Heterogeneous Causal Effects}.
}

\seealso{
  \code{\link{honest.causalTree}},
  \code{\link{causalTree}},
  \code{\link{predict.rpart}},
}

\examples{
fit <- causalTree(y~ x1 + x2 + x3 + x4, data = simulation.1, 
                  treatment = simulation.1$treatment, split.Rule = "CT", 
                  cv.option = "CT", split.Honest = T, 
                  cv.Honest = T, split.Bucket = F, xval = 5, 
                  cp = 0, minsize = 20, propensity = 0.5)
                  
opcp <- fit$cptable[,1][which.min(fit$cptable[,4])]

opfit <- prune(fit, opcp)


honestFit <- estimate.causalTree(opfit, data = simulation.2, 
                                  treatment = simulation.2$treatment)
}